feat: restructure games into dedicated classes with consistent zone pattern
BREAKING CHANGE: Reorganized game architecture for better maintainability and scalability

- **Created Individual Game Classes**:
  - CursorBeforeClickersGame.js - Dedicated class for level-based adventure game
  - CursorTextlandGame.js - Dedicated class for open exploration game
  - Each game class follows the same pattern as zone classes

- **Reorganized File Structure**:
  - Moved GameRegistry and GameProviderAdapter to src/infrastructure/data/games/
  - Created games/ subdirectory matching the existing zones/ pattern
  - Moved all game-related tests to tests/infrastructure/data/games/

- **Updated All Import Paths**:
  - Fixed 15+ import statements across the codebase
  - Updated VimForKidsGame, LevelGameState, GameFactory, services, and UI components
  - Fixed dynamic imports in PersistenceService

- **Enhanced Test Coverage**:
  - Added comprehensive tests for CursorBeforeClickersGame (17 test cases)
  - Added comprehensive tests for CursorTextlandGame (12 test cases)
  - Moved and updated existing GameRegistry and GameProviderAdapter tests

- **Benefits**:
  - Consistent architecture: Games now follow the same pattern as zones
  - Better scalability: Adding new games is as simple as creating a new class
  - Improved maintainability: Each game's configuration is self-contained
  - Easier testing: Individual game classes can be tested in isolation
  - Clear separation of concerns: Game logic separated from registry logic

- **Maintained Compatibility**: Zero breaking changes to existing functionality
- **All Tests Passing**: 937 tests passing, including new game class tests
